Add a helper to rethrow the promise
this is probably required everytime exceptions have to be handled. Not sure what other libraries do here, but this newly added rethrowPromise helper can be used to rethrow if the given error values is the promise managed by the data holer. The hook also uses this internally to propagate the promise without interpreting it as an error.